+++ lib/shared/themes/base.ts
@@ -16,13 +16,16 @@ const theme: ThemeProps = {
 	hoverColor: Color.silver,
 	inputBorderColor: Color.silver,
 	itemHoverColor: ColorScheme.c3,
+	outlineColor: Color.glow,
+	searchFocus: Color.glow,
+	searchMatch: ColorScheme.c5,
 	selectedBackgroundColor: ColorScheme.c2,
 	selectedForegroundColor: ColorScheme.c4,
 	titleBarBackgroundColor: ColorScheme.c1,
 	titleBarForegroundColor: ColorScheme.c4,
 	tooltipBackgroundColor: Color.slategray,
 	tooltipForegroundColor: Color.white,
-	transitionDelay: "0.5s"
+	transitionDelay: "0.33s"
 };
 
 export default theme;