+++ lib/textArea/index.ts
@@ -1,0 +1,14 @@
+import TextArea, {
+	getDefaultTextAreaProps,
+	getDefaultTextAreaState,
+	TextAreaProps,
+	TextAreaState
+} from "./TextArea";
+
+export {
+	getDefaultTextAreaProps,
+	getDefaultTextAreaState,
+	TextArea,
+	TextAreaProps,
+	TextAreaState
+};